+++ src/App.css
@@ -205,3 +205,11 @@ body {
   text-align: center;
   margin-top: 12px;
 }
+
+.main-img img {
+  width: 100%;
+  max-width: 300px;
+  height: auto;
+  display: block;
+  margin: 0 auto;
+}
