+++ src/pages/MainPage.js
@@ -1,20 +1,18 @@
+// MainContent.js
 import React, { useRef, useState } from 'react';
-import '../App.css';  // App의 스타일을 사용할 경우
-import { useNavigate } from 'react-router-dom';
+import { useNavigate, Link } from 'react-router-dom';
 
 import { getProjectImages } from '../utils/get-project-images'
 
 const img1 = getProjectImages(1);
 const img2 = getProjectImages(2);
 const img3 = getProjectImages(3);
+const logo = getProjectImages(4);
 
-function Home() {
+function MainPage() {
   const imgRefs = [useRef(null), useRef(null), useRef(null)];
-  const [showMenu, setShowMenu] = useState(false);
-  const [menuPinned, setMenuPinned] = useState(false);
   const [clickedIndex, setClickedIndex] = useState(null);
-
-  const navigate = useNavigate();  // useNavigate 훅 사용
+  const navigate = useNavigate();
 
   const handleMouseMove = (e, index) => {
     const card = imgRefs[index].current;
@@ -23,17 +21,13 @@ function Home() {
     const y = e.clientY - rect.top;
     const centerX = rect.width / 2;
     const centerY = rect.height / 2;
-
     const rotateX = -(y - centerY) / 6;
     const rotateY = (x - centerX) / 6;
-
     const scale = clickedIndex === index ? 1.6 : 1.1;
     const opacity = clickedIndex === index ? 0.7 : 1;
-
     const lightX = (x / rect.width) * 100;
     const lightY = (y / rect.height) * 100;
     const highlight = `radial-gradient(circle at ${lightX}% ${lightY}%, rgba(255,255,255,0.35), transparent 60%)`;
-
     card.style.transform = `scale(${scale}) rotateX(${rotateX}deg) rotateY(${rotateY}deg)`;
     card.style.opacity = opacity;
     card.style.backgroundImage = highlight;
@@ -62,12 +56,28 @@ function Home() {
       card.style.opacity = 0.7;
     }
 
-    // 이미지 클릭 시 "/photo-search" 페이지로 이동
-    navigate('/photo-search');
+    if (index === 0) navigate('/ingredient-search');
+    else if (index === 1) navigate('/photo-search');
   };
 
   return (
     <div className="app">
+      <header className="header daangn-style">
+        <div className="header-left" onClick={() => navigate('/')} style={{ cursor: 'pointer' }}>
+          <img src={logo} alt="RecipeGo" className="logo-img" />
+        </div>
+        <nav className="header-center">
+          <Link to="/ingredient-search">재료로 검색</Link>
+          <Link to="/photo-search">사진으로 검색</Link>
+          <Link to="/help">도움말</Link>
+        </nav>
+        <div className="header-right">
+          <button onClick={() => navigate('/login')} className="auth-btn">로그인</button>
+          <span className="divider">|</span>
+          <button onClick={() => navigate('/signup')} className="auth-btn">회원가입</button>
+        </div>
+      </header>
+
       <div className="main">
         {[img1, img2, img3].map((imgSrc, i) => (
           <div className="partition" key={i}>
@@ -76,7 +86,7 @@ function Home() {
               className="main-img"
               onMouseMove={(e) => handleMouseMove(e, i)}
               onMouseLeave={() => resetTransform(i)}
-              onClick={() => handleClick(i)}  // 이미지 클릭 시 페이지 이동
+              onClick={() => handleClick(i)}
               style={{ backgroundSize: 'cover', backgroundRepeat: 'no-repeat', backgroundPosition: 'center' }}
             >
               <img
@@ -97,4 +107,4 @@ function Home() {
   );
 }
 
-export default Home;
+export default MainPage;
